service_orders/app/core/security.py
from typing import Optional
from jose import JWTError, jwt
from app.config import settings


def decode_access_token(token: str) -> Optional[dict]:
    """
    Декодирует JWT access token.
    """
    try:
        payload = jwt.decode(
            token,
            settings.JWT_SECRET_KEY,
            algorithms=[settings.JWT_ALGORITHM]
        )
        return payload
    except JWTError:
        return None
